Add license headers and support REUSE 3.0
[REUSE](https://reuse.software) is a tool that specifies license terms
for each file by header, in a syntax that's digestible both for humans
and computers, and supports different licenses for different files.

using `reuse lint', we can see that every file either has a header
specifying its license or is listed in .reuse/dep5 (for autogenerated
or otherwise unannotatable files).

COPYING notice added just to make GitHub happy (it doesn't support
REUSE in the UI yet)